Refactor UserInvitationService to UserInvitationAppService and implement email invitation sending functionality
