+++ aspnet-core/test/toyiyo.todo.Tests/Invitations/UserInvitationServiceTests.cs
@@ -17,6 +17,7 @@ namespace toyiyo.todo.Tests.Invitations
         private readonly IUserInvitationAppService _userInvitationService;
         private readonly TenantManager _tenantManager;
         private readonly IEmailSender _emailSender;
+        private readonly Microsoft.Extensions.Configuration.IConfiguration _configuration;
 
         public UserInvitationServiceTests()
         {
@@ -24,6 +25,7 @@ namespace toyiyo.todo.Tests.Invitations
             _userInvitationService = Resolve<IUserInvitationAppService>();
             _tenantManager = Resolve<TenantManager>();
             _emailSender = Resolve<IEmailSender>();
+            _configuration = Resolve<Microsoft.Extensions.Configuration.IConfiguration>();
             // Call LoginAsDefaultTenantAdmin before each test
             LoginAsDefaultTenantAdmin();
         }
